+++ soc/arm/nordic_nrf/nrf51/CMakeLists.txt
@@ -1,4 +1,7 @@
 zephyr_sources(
-  power.c
   soc.c
+  )
+
+zephyr_sources_ifdef(CONFIG_SYS_POWER_MANAGEMENT
+  power.c
   )
