+++ soc/arm/nordic_nrf/nrf51/power.c
@@ -89,11 +89,3 @@ bool sys_is_valid_power_state(enum power_states state)
 
 	/* Not reached */
 }
-
-/* Overrides the weak ARM implementation:
-   Set general purpose retention register and reboot */
-void sys_arch_reboot(int type)
-{
-	nrf_power_gpregret_set((uint8_t)type);
-	NVIC_SystemReset();
-}
