+++ soc/arm/nordic_nrf/nrf52/soc_power.h
@@ -8,37 +8,16 @@
 #define _SOC_POWER_H_
 
 #include <stdbool.h>
+#include <power.h>
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-enum power_states {
-#ifdef CONFIG_SYS_POWER_LOW_POWER_STATE
-# ifdef CONFIG_SYS_POWER_STATE_CPU_LPS_SUPPORTED
-	SYS_POWER_STATE_CPU_LPS,	/* Not used */
-# endif
-# ifdef CONFIG_SYS_POWER_STATE_CPU_LPS_1_SUPPORTED
-	SYS_POWER_STATE_CPU_LPS_1,	/* Not used */
-# endif
-# ifdef CONFIG_SYS_POWER_STATE_CPU_LPS_2_SUPPORTED
-	SYS_POWER_STATE_CPU_LPS_2,	/* Not used */
-# endif
-#endif /* CONFIG_SYS_POWER_LOW_POWER_STATE */
-
-#ifdef CONFIG_SYS_POWER_DEEP_SLEEP
-# ifdef CONFIG_SYS_POWER_STATE_DEEP_SLEEP_SUPPORTED
-	SYS_POWER_STATE_DEEP_SLEEP,	/* System OFF */
-# endif
-# ifdef CONFIG_SYS_POWER_STATE_DEEP_SLEEP_1_SUPPORTED
-	SYS_POWER_STATE_DEEP_SLEEP_1,	/* Not used */
-# endif
-# ifdef CONFIG_SYS_POWER_STATE_DEEP_SLEEP_2_SUPPORTED
-	SYS_POWER_STATE_DEEP_SLEEP_2,	/* Not used */
-# endif
-#endif /* CONFIG_SYS_POWER_DEEP_SLEEP */
-	SYS_POWER_STATE_MAX		/* Do nothing */
-};
+/*
+ * Power state map:
+ * SYS_POWER_STATE_DEEP_SLEEP: System OFF
+ */
 
 /**
  * @brief Put processor into low power state
