+++ soc/x86/intel_quark/quark_se/CMakeLists.txt
@@ -12,6 +12,9 @@ zephyr_sources(
   soc.c
   soc_config.c
   eoi.c
+  )
+
+zephyr_sources_ifdef(CONFIG_SYS_POWER_MANAGEMENT
   power.c
   soc_power.S
   )
